Time Series Analysis on rainfall data. Predicting Sentiments of tweets
Tweets are placed into a corpus which is then modified, removing meaningless words that provide no valuable information. Once they are removed we check the similarity of tweets by looking at the cosine similarity of the terms in each document (or tweet). We can then use this information to predict sentiments in future tweets.